Convert FrameState struct into double-swap class
I don't think this will break much.
This also allows (later) to remove all the ref's for frame state.

In hindsight, I now worry about the performance ramifications of copying the frame state around so much. This also should probably be done for KeyboardState and MouseState, but those may be more subject to changes than the base frame state type.