Save config out of game directory.
Some platforms, such as Linux, we like install game to a non-writable path like `/usr/lib/clonedash`,
creating a `cfg` directory in it is not possible for normal user.

This patch use `<game>/cfg` if it exists for compatibility reason,
else, we save config to `Environment.SpecialFolder.ApplicationData`, which value depends OS.
See comment in source code for more info.