Adjust special folders paths
Config is not ended with `cfg` now and not under appdata.
It is `<user>\AppData\Roaming\Clone Dash` on Windows, `~/.config/Clone Dash` on Linux/macOS.

Appdata is moved to another path.
It is `<user>\AppData\Local\Clone Dash` on Windows, `~/.local/share/Clone Dash` on Linux/macOS.

Add a search path for caching something.
It is `%TEMP%\Clone Dash` or `%TMP%\Clone Dash` or `<user>\AppData\Local\Temp\Clone Dash` on Windows,
`~/.cache/Clone Dash` on Linux/macOS.
Most of the time, `%TEMP%` or `%TMP%` is `<user>\AppData\Local\Temp`, so those three paths of Windows are exactly same.